requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# heroku already has this and requires that we don't specify it
setuptools==27.2.0

# Testing
pytest==3.0.2
WebTest==2.0.23
factory-boy==2.7.0

# Management script
Flask-Script==2.0.5

# Data ingestion
beautifulsoup4==4.5.1
requests==2.12.1
